Refactor interface declaration
