+++ package.json
@@ -36,6 +36,7 @@
     "sass": "^1.53.0"
   },
   "devDependencies": {
+    "@tanstack/react-query-devtools": "^4.2.3",
     "@testing-library/dom": "^8.17.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
