+++ pages/_app.tsx
@@ -9,9 +9,11 @@ import SEO from 'seo.config';
 import { appWithTranslation } from 'next-i18next';
 import '/public/static/fonts/styles.scss';
 import 'react-h5-audio-player/src/styles.scss';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-const queryClient = new QueryClient();
-
+const queryClient = new QueryClient({
+  defaultOptions: { queries: { retry: false, refetchOnWindowFocus: false } },
+});
 export type NextPageWithLayout = NextPage & {
   getLayout?: (page: ReactElement) => ReactNode;
 };
@@ -29,6 +31,7 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
         <DefaultSeo {...SEO} />
         {getLayout(<Component {...pageProps} />)}
       </RecoilRoot>
+      <ReactQueryDevtools initialIsOpen={true} />
     </QueryClientProvider>
   );
 }
