+++ types/music.ts
@@ -1,0 +1,6 @@
+export interface MusicListTypes {
+  musicId: number;
+  musicUrl: string;
+  musicName: string;
+  musicArtist: string;
+}
